feat: Brightcove video automation and component merge
