+++ amplience-automation/automation-files/schema/brightcove-video.json
@@ -1,0 +1,5 @@
+{
+    "body": "./schemas/brightcove-video-schema.json",
+    "schemaId": "https://sfcc.com/components/brightcove-video",
+    "validationLevel": "CONTENT_TYPE"
+}
