+++ amplience-automation/automation-files/schema/schemas/all-definitions-schema.json.hbs
@@ -8,6 +8,7 @@
                 "https://sfcc.com/components/section",
                 "https://sfcc.com/components/banner",
                 "https://sfcc.com/components/banner-bynder",
+                "https://sfcc.com/components/brightcove-video",
                 "https://sfcc.com/components/curated-product",
                 "https://sfcc.com/components/card-enhanced",
                 "https://sfcc.com/components/card-list",
@@ -51,6 +52,7 @@
                 "https://sfcc.com/components/section",
                 "https://sfcc.com/components/banner",
                 "https://sfcc.com/components/banner-bynder",
+                "https://sfcc.com/components/brightcove-video",
                 "https://sfcc.com/components/curated-product",
                 "https://sfcc.com/components/personalised-container",
                 "https://sfcc.com/components/personalised-component",
@@ -84,6 +86,7 @@
                 "https://sfcc.com/components/curated-product",
                 "https://sfcc.com/components/banner",
                 "https://sfcc.com/components/banner-bynder",
+                "https://sfcc.com/components/brightcove-video",
                 "https://sfcc.com/components/personalised-container",
                 "https://sfcc.com/components/personalised-component",
                 "https://sfcc.com/components/promo",
@@ -112,6 +115,7 @@
                 "https://sfcc.com/components/curated-product",
                 "https://sfcc.com/components/banner",
                 "https://sfcc.com/components/banner-bynder",
+                "https://sfcc.com/components/brightcove-video",
                 "https://sfcc.com/components/personalised-container",
                 "https://sfcc.com/components/personalised-component",
                 "https://sfcc.com/components/promo",
@@ -291,6 +295,7 @@
                                 "https://sfcc.com/components/section",
                                 "https://sfcc.com/components/banner",
                                 "https://sfcc.com/components/banner-bynder",
+                                "https://sfcc.com/components/brightcove-video",
                                 "https://sfcc.com/components/curated-product",
                                 "https://sfcc.com/components/card-enhanced",
                                 "https://sfcc.com/components/card-list",
@@ -399,6 +404,11 @@
                                 {
                                     "id": "https://sfcc.com/components/promo-bynder",
                                     "icon": "https://bigcontent.io/cms/icons/potter/potter-promo.png",
+                                    "card": "https://amp-product.s3.eu-west-1.amazonaws.com/integration-examples/byndercards/summary-photo/index.html?headline=/bynder/name&image=/bynder/originalUrl&imageAlt=/bynder/name&src=%2F%2F\{{vse.domain}}%2Fcms%2Fcontent%2Fquery%3FfullBodyObject%3Dtrue%26query%3D%257B%2522sys.iri%2522%253A%2522http%253A%252F%252Fcontent.cms.amplience.com%252F\{{content.sys.id}}%2522%257D%26scope%3Dtree%26store%3Dstaging"
+                                },
+                                {
+                                    "id": "https://sfcc.com/components/brightcove-video",
+                                    "icon": "https://dem.a.bigcontent.io/v1/static/AnyaFinn-CTicon-BrightcoveVideo",
                                     "card": "https://amp-product.s3.eu-west-1.amazonaws.com/integration-examples/byndercards/summary-photo/index.html?headline=/bynder/name&image=/bynder/originalUrl&imageAlt=/bynder/name&src=%2F%2F\{{vse.domain}}%2Fcms%2Fcontent%2Fquery%3FfullBodyObject%3Dtrue%26query%3D%257B%2522sys.iri%2522%253A%2522http%253A%252F%252Fcontent.cms.amplience.com%252F\{{content.sys.id}}%2522%257D%26scope%3Dtree%26store%3Dstaging"
                                 }
                             ]
