+++ amplience-automation/automation-files/type/brightcove-video.json.hbs
@@ -1,0 +1,35 @@
+{
+    "contentTypeUri": "https://sfcc.com/components/brightcove-video",
+    "status": "ACTIVE",
+    "settings": {
+        "label": "Brightcove Video",
+        "icons": [
+            {
+                "size": 256,
+                "url": "https://dem.a.bigcontent.io/v1/static/AnyaFinn-CTicon-BrightcoveVideo"
+            }
+        ],
+        "visualizations": [
+            {{#each visualisations}}
+            {
+                "label": "{{this.name}}",
+                "templatedUri": "{{this.url}}/\{{locales}}/visualization?vse=\{{vse.domain}}&contentId=\{{content.sys.id}}&hub=\{{hub.name}}",
+                "default": "{{this.default}}"
+            },
+            {{/each}}
+            {
+                "label": "JSON",
+                "templatedUri": "https://delivery-preview.visualizations.content.amplience.net?id=\{{content.sys.id}}&vse=\{{vse.domain}}&locale=\{{locales}}&hub=\{{cms.hub}}&realtime=true",
+                "default": false
+            }
+        ],
+        "cards" : [ 
+            {
+                "templatedUri" : "https://bigcontent.io/cms/cards/text/index.html?headline=/videoId&src=%2F%2F\{{vse.domain}}%2Fcms%2Fcontent%2Fquery%3FfullBodyObject%3Dtrue%26query%3D%257B%2522sys.iri%2522%253A%2522http%253A%252F%252Fcontent.cms.amplience.com%252F\{{content.sys.id}}%2522%257D%26scope%3Dtree%26store%3Dstaging"
+            }
+        ]
+    },
+    "repositories": [
+        "content"
+    ]
+}
