+++ app/components/amplience/wrapper/index.jsx
@@ -21,7 +21,7 @@ import {useLocation} from 'react-router-dom'
 import ShoppableVideo from '../shoppable-video'
 import Features from '../features'
 import GridLock from '../grid-lock'
-import Brightcove from '../brightcove'
+import BrightcoveVideo from '../brightcove-video'
 import FormBuilder from '../form-builder'
 import Stylitics from '../stylitics'
 import Banner from '../banner'
@@ -55,7 +55,7 @@ const componentsMapping = {
     'https://sfcc.com/components/card-list': CardList,
     'https://sfcc.com/components/features': Features,
     'https://sfcc.com/components/grid-lock': GridLock,
-    'https://sfcc.com/components/video': Brightcove,
+    'https://sfcc.com/components/brightcove-video': BrightcoveVideo,
     'https://sfcc.com/components/form-builder': FormBuilder,
 
     'https://sfcc.com/content/stylitics/generic': Stylitics,
